+++ bin/setup.sh
@@ -1365,6 +1365,7 @@ Description=Mysterium Node Monitoring Toolkit
 After=${_AFTER_DEPS}
 Wants=network-online.target
 StartLimitIntervalSec=0
+StartLimitBurst=0
 
 [Service]
 Type=simple